% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_factor.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Convert column to factor}
\usage{
convert_to_factor(data, column)
}
\arguments{
\item{data}{A data frame containing the column to be converted.}

\item{column}{The name of the column to convert (as a string).}
}
\value{
A data frame with the specified column converted to a factor.
}
\description{
This function converts a column with textual values into a factor variable, which is useful for categorical data analyses.
}
\examples{
data <- data.frame(gender = c("male", "female", "female", "male"))
convert_to_factor(data, "gender")
}
