% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_datetime.R
\name{split_datetime}
\alias{split_datetime}
\title{Split datetime column into separate date and time columns}
\usage{
split_datetime(datetime_column)

split_datetime(datetime_column)
}
\arguments{
\item{datetime_column}{A vector of datetime values (POSIXct).}
}
\value{
A data.frame with separate date and time columns.

A data.frame with separate date and time columns.
}
\description{
This function takes a datetime column and splits it into separate date and time columns.

This function takes a column of datetime values and splits it into separate date and time columns.
}
\examples{
datetime_example <- as.POSIXct(c("2020-01-01 12:00:00", "2020-01-02 13:30:00"))
split_datetime(datetime_example)
data <- data.frame(datetime = as.POSIXct(c("2020-01-01 12:00:00", "2020-01-02 13:30:00")))
split_datetime(data$datetime)
}
